% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{desaturate}
\alias{desaturate}
\title{Decrease the saturation of a color}
\usage{
desaturate(color, prop)
}
\arguments{
\item{color}{A matplotlib-compatible color.}

\item{prop}{Proportion (in \verb{[0, 1]}) of the original saturation to keep.}
}
\value{
A hex color string.
}
\description{
Port of \code{seaborn.desaturate}.
}
