% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscplot.R
\name{dogplot}
\alias{dogplot}
\title{Who's a good boy?}
\usage{
dogplot(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Port of \code{seaborn.dogplot} (an easter egg). Prints an affirmation.
}
