% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscplot.R
\name{palplot}
\alias{palplot}
\title{Plot the values in a color palette as a horizontal array}
\usage{
palplot(pal, size = 1)
}
\arguments{
\item{pal}{A sequence of colors (e.g. from \link{color_palette}).}

\item{size}{Scaling factor for the swatch size.}
}
\value{
A \code{reaborn_plot}.
}
\description{
Port of \code{seaborn.palplot}. Returns a \link{reaborn_plot}.
}
\examples{
palplot(color_palette("deep"))
palplot(color_palette("rocket", 8))
}
