% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaborn-package.R
\docType{package}
\name{reaborn-package}
\alias{reaborn}
\alias{reaborn-package}
\title{reaborn: Statistical Data Visualization, the Seaborn Way}
\description{
reaborn is an R port of the Python seaborn library. It mirrors seaborn's
public function API (identical names, arguments, and defaults) and produces
visually indistinguishable plots built on ggplot2. Every reaborn plot is a
ggplot object, so it composes with the full ggplot2 grammar of graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shawnschwartz/reaborn}
  \item Report bugs at \url{https://github.com/shawnschwartz/reaborn/issues}
}

}
\author{
\strong{Maintainer}: Shawn Schwartz \email{shawn.t.schwartz@gmail.com}

Authors:
\itemize{
  \item Shawn Schwartz \email{shawn.t.schwartz@gmail.com}
}

}
\keyword{internal}
