% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmod-theme.R
\name{reset_defaults}
\alias{reset_defaults}
\alias{reset_orig}
\title{Restore matplotlib/ggplot2 defaults}
\usage{
reset_defaults()

reset_orig()
}
\value{
Invisibly \code{NULL}.
}
\description{
Port of \code{seaborn.reset_defaults} / \code{seaborn.reset_orig}.
}
