% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{saturate}
\alias{saturate}
\title{Increase the saturation of a color to its maximum}
\usage{
saturate(color)
}
\arguments{
\item{color}{A matplotlib-compatible color.}
}
\value{
A hex color string.
}
\description{
Port of \code{seaborn.saturate}.
}
