% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{set_hls_values}
\alias{set_hls_values}
\title{Independently set the hue, lightness, and/or saturation of a color}
\usage{
set_hls_values(color, h = NULL, l = NULL, s = NULL)
}
\arguments{
\item{color}{A matplotlib-compatible color.}

\item{h, l, s}{New hue, lightness, saturation in \verb{[0, 1]}, or \code{NULL} to keep.}
}
\value{
A hex color string.
}
\description{
Port of \code{seaborn.set_hls_values}.
}
