% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmod-theme.R
\name{set_palette}
\alias{set_palette}
\title{Set the matplotlib color cycle / ggplot default discrete palette}
\usage{
set_palette(palette, n_colors = NULL, desat = NULL, color_codes = FALSE)
}
\arguments{
\item{palette}{A palette name or sequence (see \link{color_palette}).}

\item{n_colors, desat, color_codes}{See seaborn.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Port of \code{seaborn.set_palette}.
}
