% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmod-theme.R
\name{set_theme}
\alias{set_theme}
\alias{set}
\title{Set multiple theme parameters in one step}
\usage{
set_theme(
  context = "notebook",
  style = "darkgrid",
  palette = "deep",
  font = "sans",
  font_scale = 1,
  color_codes = TRUE,
  rc = NULL
)

set(...)
}
\arguments{
\item{context, style, palette, font, font_scale, color_codes, rc}{See seaborn.}

\item{...}{Passed to \link{set_theme}.}
}
\value{
Invisibly, the applied \link[ggplot2:theme]{ggplot2::theme}.
}
\description{
Port of \code{seaborn.set_theme} (and its alias \code{set}). Sets the global look used
by subsequent reaborn (and ggplot2) plots.
}
