% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmod-theme.R
\name{theme_seaborn}
\alias{theme_seaborn}
\title{Build a ggplot2 theme replicating a seaborn style + context}
\usage{
theme_seaborn(
  style = "darkgrid",
  context = "notebook",
  font_scale = 1,
  font = "sans"
)
}
\arguments{
\item{style}{A seaborn style name (see \link{axes_style}).}

\item{context}{A seaborn context name (see \link{plotting_context}).}

\item{font_scale}{Font scaling factor.}

\item{font}{Base font family.}
}
\value{
A complete \link[ggplot2:theme]{ggplot2::theme} object.
}
\description{
Build a ggplot2 theme replicating a seaborn style + context
}
